Refactor GliaViewController environment initialization
