+++ GliaWidgets/Sources/ViewController/GliaViewController.swift
@@ -175,20 +175,3 @@ extension GliaViewController: UIViewControllerTransitioningDelegate {
         )
     }
 }
-
-extension GliaViewController {
-    struct Environment {
-        var uiApplication: UIKitBased.UIApplication
-        var uiScreen: UIKitBased.UIScreen
-        var log: CoreSdkClient.Logger
-        var animate: (_ animated: Bool, _ animations: @escaping () -> Void, _ completion: @escaping (Bool) -> Void) -> Void
-
-        func withAnimation(
-            animated: Bool,
-            animations: @escaping () -> Void,
-            completion: @escaping (Bool) -> Void
-        ) {
-            animate(animated, animations, completion)
-        }
-    }
-}
